requirements.txt
opencv-python
mediapipe
numpy
streamlit
pyyaml
pytest